+++ ObjDetector/LearningAlgorithms/CascadeClassifier/StageClassifiers/StageClassifier.cpp
@@ -7,36 +7,36 @@ using std::count_if;
 using std::bind2nd;
 using std::greater_equal;
 
-void StageClassifier::SearchTheta(vector<vector<vector<double>>>& X, vector<bool>& y)
+void StageClassifier::SearchTheta(vector<vector<vector<float>>>& X, vector<bool>& y)
 {
     int whole_n_total = (int)y.size();
     int whole_n_pos = (int)count(y.begin(), y.end(), true);
     int whole_n_neg = (int)(whole_n_total - whole_n_pos);
 
-    vector<double> probs;
+    vector<float> probs;
 
     for (int i = 0; i < X.size(); i++)
         probs.push_back(Predict(X[i]));
 
     /* search min TPR */
-    double threshhold;
+    float threshhold;
     for (threshhold = 1; threshhold >= 0; threshhold -= search_step)
     {
-        TPR = count_if(probs.begin(), probs.begin() + whole_n_pos, bind2nd(greater_equal<double>(), threshhold)) / (double)whole_n_pos;
+        TPR = count_if(probs.begin(), probs.begin() + whole_n_pos, bind2nd(greater_equal<float>(), threshhold)) / (float)whole_n_pos;
         if (TPR >= TPR_min)
             break;
     }
 
     /* get corresponding theta and FPR */
     theta = threshhold;
-    FPR = count_if(probs.begin() + whole_n_pos, probs.end(), bind2nd(greater_equal<double>(), threshhold)) / (double)whole_n_neg;
+    FPR = count_if(probs.begin() + whole_n_pos, probs.end(), bind2nd(greater_equal<float>(), threshhold)) / (float)whole_n_neg;
 }
 
-double StageClassifier::Evaluate(vector<vector<vector<double>>>& X, vector<bool>& y)
+float StageClassifier::Evaluate(vector<vector<vector<float>>>& X, vector<bool>& y)
 {
-    vector<double> probs;
-    vector<double> TPRs, FPRs;
-    double area = 0;
+    vector<float> probs;
+    vector<float> TPRs, FPRs;
+    float area = 0;
 
     for (int i = 0; i < X.size(); i++)
         probs.push_back(Predict(X[i]));
@@ -53,13 +53,13 @@ double StageClassifier::Evaluate(vector<vector<vector<double>>>& X, vector<bool>
     //}
     //LOG_DEBUG_NN("\t\tStrong classifier: ");
     //LOG_DEBUG_NN("TP = " << TP << '/' << y.size() / 2 << ", TN = " << TN << '/' << y.size() / 2);
-    //LOG_DEBUG_NN(", Result: " << (double)(TP + TN) / y.size());
+    //LOG_DEBUG_NN(", Result: " << (float)(TP + TN) / y.size());
     //#endif
 
-    for (double threshhold = 1; threshhold >= 0; threshhold -= auc_step)
+    for (float threshhold = 1; threshhold >= 0; threshhold -= auc_step)
     {
-        TPRs.push_back(count_if(probs.begin(), probs.begin() + n_pos, bind2nd(greater_equal<double>(), threshhold)) / (double)n_pos);
-        FPRs.push_back(count_if(probs.begin() + n_pos, probs.end(), bind2nd(greater_equal<double>(), threshhold)) / (double)n_neg);
+        TPRs.push_back(count_if(probs.begin(), probs.begin() + n_pos, bind2nd(greater_equal<float>(), threshhold)) / (float)n_pos);
+        FPRs.push_back(count_if(probs.begin() + n_pos, probs.end(), bind2nd(greater_equal<float>(), threshhold)) / (float)n_neg);
 
         if (FPRs.size() > 1)
             area += TPRs.back() * (FPRs.back() - FPRs[FPRs.size() - 2]); // right Riemann sum
